+++ config/options.js
@@ -8,7 +8,7 @@ const generatorNames = Object.keys(generators);
 
 const generalOptions = [
   {
-    param: '-D, --debug ',
+    param: '-d, --debug ',
     description: 'debug',
   },
   {
@@ -16,14 +16,8 @@ const generalOptions = [
     description: 'Skip setup questions',
   },
   {
-    param: '-f, --frontend <frontend>',
-    description: 'Frontend framework (React or Angular)',
-    validation: /^(react|angular)$/i,
-    fallback: 'react',
-  },
-  {
     param: '-b, --backend <backend>',
-    description: 'Backend framework (Node.js, .NET Core or none)',
+    description: 'Backend framework (Node.js or none)',
     validation: /^(nodejs|dotnet)$/i,
     fallback: 'nodejs',
   },