+++ config/questions.js
@@ -13,15 +13,6 @@ const questions = [
       { value: undefined, name: 'I don\'t need a BFF' },
     ],
   },
-  {
-    type: 'list',
-    name: 'frontend',
-    message: 'Which frontend framework would you like?',
-    choices: [
-      { value: 'angular', name: 'Angular' },
-      { value: 'react', name: 'React' },
-    ],
-  },
 ];
 
 module.exports = questions;