+++ utils/delete.js
@@ -7,4 +7,31 @@ function deleteFolderSync(folder) {
   }
 }
 
-module.exports.deleteFolderSync = deleteFolderSync;
+function deleteFolderRecursive(folderpath) {
+  if (fs.existsSync(folderpath)) {
+    fs.readdirSync(folderpath).forEach((file) => {
+      const curPath = `${folderpath}/${file}`;
+      if (fs.lstatSync(curPath).isDirectory()) { // recurse
+        deleteFolderRecursive(curPath);
+      } else { // delete file
+        fs.unlinkSync(curPath);
+      }
+    });
+    fs.rmdirSync(folderpath);
+  }
+}
+
+function deleteFile(filepath) {
+  try {
+    return fs.unlinkSync(filepath);
+  } catch (e) {
+    console.log('deleteFile error:', e);
+    throw e;
+  }
+}
+
+module.exports = {
+  deleteFolderSync,
+  deleteFolderRecursive,
+  deleteFile,
+};